move pilot functions to own module
