+++ ol/cli/src/config.rs
@@ -1,3 +1,2 @@
 //! `config` module for all 0L settings
-pub use ol_types
-::config::AppCfg;
+pub use ol_types::config::AppCfg;
