+++ ol/cli/src/mgmt/management.rs
@@ -32,23 +32,6 @@ pub struct HostProcess {
     pids: HashSet<u32>,
 }
 
-/// Check if we are in prod mode
-// pub static IS_PROD: Lazy<bool> = Lazy::new(|| {
-//     match env::var("NODE_ENV") {
-//         Ok(val) => {
-//             match val.as_str() {
-//                 "prod" => true,
-//                 // if anything else is set by user is false
-//                 _ => false,
-//             }
-//         }
-//         // default to prod if nothig is set
-//         _ => true,
-//     }
-// });
-
-// TODO: do we need to kill zombies this way?
-
 /// create log files
 pub fn create_log_file(file_name: &str) -> File {
     let conf = app_config();
@@ -78,45 +61,6 @@ fn spawn_process(
         .spawn()
         .expect(expect_msg)
 }
-
-// /// start validator wizard
-// pub fn run_validator_wizard() -> bool {
-//     println!("Running validator wizard");
-//     let entry_arg = entrypoint::get_args();
-
-//     let mut child = if *IS_PROD {
-//         Command::new("miner")
-//             .arg("val-wizard")
-//             .spawn()
-//             .expect(&format!("failed to find 'miner', is it installed?"))
-//     } else if let Some(path) = entry_arg.swarm_path {
-//         // we are testing with swarm
-//         let swarm_arg = path.to_str().unwrap();
-//         let swarm_persona = entry_arg.swarm_persona.unwrap();
-
-//         Command::new("cargo")
-//             .args(&["r", "-p", "miner", "--"])
-//             .arg("--swarm-path")
-//             .arg(swarm_arg)
-//             .arg("--swarm-persona")
-//             .arg(swarm_persona)
-//             .arg("val-wizard")
-//             .spawn()
-//             .expect(&format!("failed to run cargo r -p miner"))
-//     } else {
-//         // we are testing on devnet
-//         Command::new("cargo")
-//             .args(&["r", "-p", "miner", "--"])
-//             .arg("val-wizard")
-//             .spawn()
-//             .expect(&format!("failed to run cargo r -p miner"))
-//     };
-
-//     let exit_code = child.wait().expect("failed to wait on miner");
-//     assert!(exit_code.success());
-
-//     true
-// }
 
 impl Node {
     /// Start Node, as fullnode
